chore(notebooks): move preview helpers to notebook utils
