+++ frontend/src/scenes/notebooks/Notebook/Editor.tsx
@@ -57,8 +57,7 @@ import { NotebookNodeZendeskTickets } from '../Nodes/NotebookNodeZendeskTickets'
 import { FloatingSuggestions } from '../Suggestions/FloatingSuggestions'
 import { insertionSuggestionsLogic } from '../Suggestions/insertionSuggestionsLogic'
 import { NotebookEditor } from '../types'
-import { textContent } from '../utils'
-import { notebookEditorLogicKey, shouldUseNotebookCollab } from './collabUtils'
+import { notebookEditorLogicKey, shouldUseNotebookCollab, textContent } from '../utils'
 import { CollapsibleHeading } from './CollapsibleHeading'
 import { DropAndPasteHandlerExtension } from './DropAndPasteHandlerExtension'
 import { InlineMenu } from './InlineMenu'
