+++ frontend/src/scenes/notebooks/Notebook/notebookCollabLogic.test.ts
@@ -4,6 +4,7 @@ import { EditorState } from '@tiptap/pm/state'
 
 import { TTEditor } from 'lib/components/RichContentEditor/types'
 
+import { notebookEditorLogicKey, shouldUseNotebookCollab } from '../utils'
 import { applyRemoteStep, RemoteStep } from './notebookCollabLogic'
 import { REMOTE_PRESENCE_META } from './RemotePresenceExtension'
 
@@ -48,6 +49,27 @@ function presence(head: number): RemoteStep['presence'] {
 }
 
 describe('notebookCollabLogic', () => {
+    describe('shouldUseNotebookCollab', () => {
+        it.each([
+            { collabEnabled: true, hasNotebook: true, hasPreviewContent: false, expected: true },
+            { collabEnabled: false, hasNotebook: true, hasPreviewContent: false, expected: false },
+            { collabEnabled: true, hasNotebook: false, hasPreviewContent: false, expected: false },
+            { collabEnabled: true, hasNotebook: true, hasPreviewContent: true, expected: false },
+        ])(
+            'returns $expected for collabEnabled=$collabEnabled, hasNotebook=$hasNotebook, hasPreviewContent=$hasPreviewContent',
+            ({ collabEnabled, hasNotebook, hasPreviewContent, expected }) => {
+                expect(shouldUseNotebookCollab(collabEnabled, hasNotebook, hasPreviewContent)).toBe(expected)
+            }
+        )
+    })
+
+    describe('notebookEditorLogicKey', () => {
+        it('uses a different editor key when collaboration is active', () => {
+            expect(notebookEditorLogicKey('abc123', true)).toBe('Notebook.abc123-collab')
+            expect(notebookEditorLogicKey('abc123', false)).toBe('Notebook.abc123')
+        })
+    })
+
     describe('applyRemoteStep', () => {
         it('applies a step when remote.version matches local + 1', () => {
             const editor = createMockEditor(0)
