+++ frontend/src/scenes/notebooks/utils.ts
@@ -113,3 +113,15 @@ export function updateContentHeading(content: JSONContent, newTitle: string): JS
         ],
     }
 }
+
+export function shouldUseNotebookCollab(
+    collabEnabled: boolean,
+    hasNotebook: boolean,
+    hasPreviewContent: boolean
+): boolean {
+    return collabEnabled && hasNotebook && !hasPreviewContent
+}
+
+export function notebookEditorLogicKey(shortId: string, useCollab: boolean): string {
+    return `Notebook.${shortId}${useCollab ? '-collab' : ''}`
+}
